11th commit

+++ .vscode/Q_15964/Q_15964.c
@@ -4,11 +4,11 @@
 
 int main(void)
 {
-    int A, B;
+    long long A, B;
 
-    scanf("%d %d", &A, &B);
+    scanf("%lld %lld", &A, &B);
 
-    printf("%d\n", (A+B)*(A-B));
-    
+    printf("%lld\n", (A+B)*(A-B));
+
     return 0;
 }
